data/generate_data/requirements.txt
ortools
cryptography==41.0.7
secrets
shapely
argparse
matplotlib
pyyaml